xplat: Reduce Static/Shared Library differences.
This will greatly reduce time required to rebuild and switch between
static and shared libraries.

Note: for Xcode/IOS, the force_load parameter is only good for the
next library mentioned. This patch forces all of the existing
libraries to get loaded into a single library libChakraCoreStatic,
and hopefully also will make it to where libChakraCore has all the
required symbols too.
