+++ bin/CMakeLists.txt
@@ -1,5 +1,4 @@
+add_subdirectory (ChakraCore)
 add_subdirectory (GCStress)
 add_subdirectory (ch)
-if(NOT STATIC_LIBRARY)
-    add_subdirectory (ChakraCore)
-endif()
+
