+++ bin/GCStress/StubExternalApi.cpp
@@ -111,3 +111,9 @@ HRESULT MemProtectHeapSynchronizeWithCollector(void * heapHandle) { return E_NOT
 #if DBG && defined(INTERNAL_MEM_PROTECT_HEAP_ALLOC)
 void MemProtectHeapSetDisableConcurrentThreadExitedCheck(void * heapHandle) {};
 #endif
+
+
+_NOINLINE void VALIDATE_ENTER_CURRENT_THREAD()
+{
+     // Does nothing
+}
