+++ lib/Jsrt/jsrtHelper.h
@@ -15,9 +15,6 @@ private:
     JsrtContext* originalJsrtContext;
 };
 
-#if defined(CHAKRA_STATIC_LIBRARY) || !defined(_WIN32)
+
 _NOINLINE void VALIDATE_ENTER_CURRENT_THREAD();
-#else
-// Windows  Shared Library: DllMain is responsible from handling all these stuff
-#define VALIDATE_ENTER_CURRENT_THREAD()
-#endif
+
